requirements.txt
alembic==1.1.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Admin==1.5.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
typed-ast==1.4.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.5
wrapt==1.11.2
WTForms==2.2.1
